Move reusable stuff from sendFlightRemarksNotification into utils file
